+++ src/main/java/com/ebay/llm/qos/config/Model.java
@@ -7,8 +7,8 @@ public class Model {
 
   private String id;
   private String description;
-  private Integer apiLimitPerMinute;
-  private Integer tokensLimitPerRequest;
-  private Integer tokensLimitPerMinute;
-  private Integer coolingPeriodSeconds;
+  private Long apiLimitPerMinute;
+  private Long tokensLimitPerRequest;
+  private Long tokensLimitPerMinute;
+  private Long coolingPeriodSeconds;
 }
